requirements.txt
asgiref==3.4.1
Django==3.2.7
django-phonenumber-field==5.2.0
phonenumbers==8.12.34
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
